updatedeps: moved class  in its own file
