+++ CHANGELOG.md
@@ -15,3 +15,4 @@
 - feature: added class `$String` for everybody who is facing issues when `String` is used.
 - feature: `String.isNullOrWhiteSpace` accepts null or undefined now
 - chore: Updated packages
+- chore: added linting
