+++ README.md
@@ -1,7 +1,10 @@
 
+[![CircleCI](https://circleci.com/gh/sevensc/typescript-string-operations.svg?style=shield)](https://app.circleci.com/pipelines/github/sevensc/typescript-string-operations)
+[![Quality Gate Status](https://sonarcloud.io/api/project_badges/measure?project=sevensc_typescript-string-operations&metric=alert_status)](https://sonarcloud.io/dashboard?id=sevensc_typescript-string-operations)
+![npm](https://img.shields.io/npm/v/typescript-string-operations)
+![npm](https://img.shields.io/npm/dw/typescript-string-operations)
 
-![CircleCI](https://img.shields.io/circleci/build/github/iwt-svenulrich/typescript-string-operations?logo=circleci&token=9234d9f6803b37ebfcd4887fa2d6d51aa2cf5214)
-[![Quality Gate Status](https://sonarcloud.io/api/project_badges/measure?project=sevensc_typescript-string-operations&metric=alert_status)](https://sonarcloud.io/dashboard?id=sevensc_typescript-string-operations)![npm](https://img.shields.io/npm/v/typescript-string-operations)![npm](https://img.shields.io/npm/dw/typescript-string-operations)
+
 # Simple lightweight string operation library for Typescript.
 ## No jQuery required! Unit tested, works with Angular.
 
@@ -9,47 +12,53 @@
  import { String, StringBuilder } from 'typescript-string-operations';
  ```
 
+I recently got feedback that sometimes there are issues when using `String`, since its actually replacing the native `String` object from JavaScript. Youcan use `$String` instead. 
+
+```typescript
+ import { $String } from 'typescript-string-operations';
+```
+
 #### USAGE:
 
-### String.Empty
+### String.empty
 ```typescript
-var id = String.Empty;
+var id = String.empty;
 ```
 
-### String.IsNullOrWhiteSpace():
+### String.isNullOrWhiteSpace():
 ```typescript
 var id = image.GetId();
-if(String.IsNullOrWhiteSpace(id))
+if(String.isNullOrWhiteSpace(id))
 	return image;
 ```
-### String.Format():
+### String.format():
 
 ```typescript
 var id = image.GetId()
-String.Format("image_{0}.jpg", id)
+String.format("image_{0}.jpg", id)
 output: "image_2db5da20-1c5d-4f1a-8fd4-b41e34c8c5b5.jpg";
 ```
 
 Specifier available!
 ```typescript
-var value = String.Format("{0:L}", "APPLE"); //output "apple"
+var value = String.format("{0:L}", "APPLE"); //output "apple"
 
-value = String.Format("{0:U}", "apple"); // output "APPLE"
+value = String.format("{0:U}", "apple"); // output "APPLE"
 
-value = String.Format("{0:d}", "2017-01-23 00:00"); //output "23.01.2017"
+value = String.format("{0:d}", "2017-01-23 00:00"); //output "23.01.2017"
 
 
-value = String.Format("{0:s}", "21.03.2017 22:15:01") //output "2017-03-21T22:15:01"
+value = String.format("{0:s}", "21.03.2017 22:15:01") //output "2017-03-21T22:15:01"
 
-value = String.Format("{0:n}", 1000000);
+value = String.format("{0:n}", 1000000);
 //output "1.000.000"
 
-value = String.Format("{0:00}", 1);
+value = String.format("{0:00}", 1);
 //output "01"
 ```
 
 ## UPDATE
-#### String Format for Objects including specifiers
+#### String.format for Objects including specifiers
 
 ```typescript
 var fruit = new Fruit();
@@ -58,7 +67,7 @@ fruit.color = "RED";
 fruit.shippingDate = new Date(2018, 1, 1);
 fruit.amount = 10000;
 
-String.Format("the {type:U} is {color:L} shipped on {shippingDate:s} with an amount of {amount:n}", fruit);
+String.format("the {type:U} is {color:L} shipped on {shippingDate:s} with an amount of {amount:n}", fruit);
 // output: the APPLE is red shipped on 2018-01-01 with an amount of 10.000
 
 ```
@@ -78,18 +87,18 @@ String.Format("the {type:U} is {color:L} shipped on {shippingDate:s} with an amo
 ### String.Join():
 
 ```typescript
-var value = String.Join("; ", "Apple", "Banana");
+var value = String.join("; ", "Apple", "Banana");
 //output: "Apple; Banana";
 ```
 #### OR
 
 ```typescript
  let object = { Name: "Foo", Value: "Bar" };
- var value = String.Join('.', object);
+ var value = String.join('.', object);
 //output: "Foo.Bar";
 
 var array = ['Apple', 'Banana']
-var value = String.Join("; ", array);
+var value = String.join("; ", array);
 //output: "Apple; Banana";
 ```
 
@@ -97,11 +106,11 @@ var value = String.Join("; ", array);
 
 | Method                    |  Type       |       Description          | Parameter  |
 | :------------------------:|:-----------:|:--------------------------:|:----------:|
-|  `Empty`                  | `Property`  |    simply returns `""`.    |
-| `IsNullOrWhiteSpace`      | `Method`    | returns true value if given parameter is either null, empty or undefined. | `format`, `args`
-| `Format`                  | `Method`    | Converts the value of objects to strings based on the formats specified and inserts them into another string. | `format`, `args`
-| `Join`                    | `Method`    |   Combines arguments delimited by given seperator.| `delimiter`,`args`
-| `Join`                    | `Method`    |   Combines arguments delimited by given seperator from array. | `delimiter`,`array` |
+|  `empty`                  | `Property`  |    simply returns `""`.    |
+| `isNullOrWhiteSpace`      | `Method`    | returns true value if given parameter is either null, empty or undefined. | `format`, `args`
+| `format`                  | `Method`    | Converts the value of objects to strings based on the formats specified and inserts them into another string. | `format`, `args`
+| `join`                    | `Method`    |   Combines arguments delimited by given seperator.| `delimiter`,`args`
+| `join`                    | `Method`    |   Combines arguments delimited by given seperator from array. | `delimiter`,`array` |
 
 
 ### StringBuilder
