+++ tsconfig.json
@@ -5,7 +5,7 @@
     "compilerOptions": {
         "noImplicitAny": true,
         "removeComments": true,
-        "target": "ES2022", 
+        "target": "ES5", 
         "module": "commonjs",
         "outDir": "./dist",
         "skipLibCheck": true,
